Fixed a bug
The bug was that, if you only entered one number and an operation and the clicked calc, the program would crash with a IndexError. So, I caught it so that it now returns an "illegal input" to be displayed on the calculator's screen. Also I changed the name of main.exe to calculator.exe to make it more easier for any non-programmer users.